Refactor budgets and usage dashboard UI
