+++ frontend/src/components/features/admin-dashboard/usage-dashboard.tsx
@@ -1,6 +1,5 @@
 /* eslint-disable i18next/no-literal-string */
 import React, { useMemo, useState } from "react";
-import { ConfirmationModal } from "#/components/shared/modals/confirmation-modal";
 import { useSelectedOrganizationId } from "#/context/use-selected-organization";
 import { useStopConversation } from "#/hooks/mutation/use-stop-conversation";
 import { useOrgConversationStats } from "#/hooks/query/use-org-conversation-stats";
@@ -9,376 +8,21 @@ import { useOrgUsageStats } from "#/hooks/query/use-org-usage-stats";
 import { useOrgUserUsage } from "#/hooks/query/use-org-user-usage";
 import { useOrganizations } from "#/hooks/query/use-organizations";
 import { organizationService } from "#/api/organization-service/organization-service.api";
-
-// Icons as inline SVGs
-function SearchIcon() {
-  return (
-    <svg
-      width="16"
-      height="16"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-    >
-      <circle cx="11" cy="11" r="8" />
-      <path d="m21 21-4.35-4.35" />
-    </svg>
-  );
-}
-
-function ExportIcon() {
-  return (
-    <svg
-      width="14"
-      height="14"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-    >
-      <path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4" />
-      <polyline points="7,10 12,15 17,10" />
-      <line x1="12" y1="15" x2="12" y2="3" />
-    </svg>
-  );
-}
-
-function StopIcon() {
-  return (
-    <svg
-      width="14"
-      height="14"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-    >
-      <circle cx="12" cy="12" r="10" />
-      <rect x="9" y="9" width="6" height="6" />
-    </svg>
-  );
-}
-
-function TrendUpIcon() {
-  return (
-    <svg
-      width="12"
-      height="12"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-    >
-      <polyline points="23,6 13.5,15.5 8.5,10.5 1,18" />
-      <polyline points="17,6 23,6 23,12" />
-    </svg>
-  );
-}
-
-function TrendDownIcon() {
-  return (
-    <svg
-      width="12"
-      height="12"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-    >
-      <polyline points="23,18 13.5,8.5 8.5,13.5 1,6" />
-      <polyline points="17,18 23,18 23,12" />
-    </svg>
-  );
-}
-
-// Format tokens
-const formatTokens = (tokens: number) => {
-  if (tokens >= 1000000) return `${(tokens / 1000000).toFixed(1)}M`;
-  if (tokens >= 1000) return `${(tokens / 1000).toFixed(1)}K`;
-  return tokens.toString();
-};
-
-// Format cost
-const formatCost = (cost: number) => {
-  if (cost >= 1000) return `$${(cost / 1000).toFixed(1)}k`;
-  return `$${cost.toFixed(2)}`;
-};
-
-// Format short date
-const formatShortDate = (dateStr: string) => {
-  const date = new Date(dateStr);
-  return date.toLocaleDateString("en-US", {
-    month: "short",
-    day: "numeric",
-  });
-};
-
-// Time window options
-const TIME_WINDOWS = [
-  { label: "7d", value: "7d" },
-  { label: "30d", value: "30d" },
-  { label: "90d", value: "90d" },
-  { label: "YTD", value: "ytd" },
-];
-
-const AGENT_COLORS = [
-  "#3B82F6",
-  "#F59E0B",
-  "#10B981",
-  "#8B5CF6",
-  "#EF4444",
-  "#06B6D4",
-  "#F97316",
-];
-
-// Format date/time
-const formatDateTime = (dateStr: string) => {
-  const date = new Date(dateStr);
-  return date.toLocaleDateString("en-US", {
-    month: "short",
-    day: "numeric",
-    year: "numeric",
-    hour: "numeric",
-    minute: "2-digit",
-  });
-};
-
-const formatDateTimeOrDash = (value?: string | null) =>
-  value ? formatDateTime(value) : "-";
-
-const formatDuration = (start?: string | null, end?: string | null) => {
-  if (!start || !end) return "-";
-  const startMs = new Date(start).getTime();
-  const endMs = new Date(end).getTime();
-  if (Number.isNaN(startMs) || Number.isNaN(endMs)) return "-";
-  const diffMs = Math.max(0, endMs - startMs);
-  const totalMinutes = Math.floor(diffMs / 60000);
-  const totalHours = Math.floor(totalMinutes / 60);
-  if (totalHours >= 24) {
-    const days = Math.floor(totalHours / 24);
-    const hours = totalHours % 24;
-    return `${days}d ${hours}h`;
-  }
-  if (totalHours > 0) {
-    return `${totalHours}h ${totalMinutes % 60}m`;
-  }
-  return `${totalMinutes}m`;
-};
-
-const formatAssociatedPr = (conversation: {
-  pr_number?: number[];
-  selected_repository?: string | null;
-}) => {
-  const prNumbers = conversation.pr_number ?? [];
-  if (prNumbers.length === 0) return "-";
-  const repo = conversation.selected_repository;
-  return prNumbers.map((pr) => (repo ? `${repo}#${pr}` : `#${pr}`)).join(", ");
-};
-
-const formatBudget = (user: {
-  budget_monthly_limit?: number | null;
-  budget_is_disabled?: boolean;
-}) => {
-  if (user.budget_is_disabled) return "Disabled";
-  if (user.budget_monthly_limit == null) return "-";
-  return formatCost(user.budget_monthly_limit);
-};
-
-const formatAgentLabel = (conversation: {
-  agent_kind?: string | null;
-  llm_model?: string | null;
-}) => {
-  const agentKind = conversation.agent_kind ?? null;
-  if (agentKind === "acp") {
-    const llmModel = conversation.llm_model ?? "";
-    const llmModelLower = llmModel.toLowerCase();
-    if (!llmModel) return "ACP";
-    if (llmModelLower.includes("claude")) return "Claude";
-    if (llmModelLower.includes("codex")) return "Codex";
-    if (llmModelLower.includes("gpt") || llmModelLower.includes("openai")) {
-      return "OpenAI";
-    }
-    if (llmModelLower.includes("gemini")) return "Gemini";
-    return llmModel;
-  }
-  if (agentKind === "openhands") return "OpenHands";
-  return agentKind || "-";
-};
-
-const formatMergedStatus = (merged?: boolean | null) => {
-  if (merged === true) return "Yes";
-  if (merged === false) return "No";
-  return "-";
-};
+import {
+  ConversationsTab,
+  ModelsTab,
+  OverviewTab,
+  UsersTab,
+} from "./usage-dashboard-tabs";
+import {
+  AGENT_COLORS,
+  TIME_WINDOWS,
+  formatCost,
+} from "./usage-dashboard-utils";
 
 // Tabs
 const TABS = ["overview", "users", "models", "conversations"] as const;
 type TabType = (typeof TABS)[number];
-
-// KPI Card component
-function KPICard({
-  label,
-  value,
-  trend,
-  trendUp,
-}: {
-  label: string;
-  value: string | number;
-  trend?: string;
-  trendUp?: boolean;
-}) {
-  return (
-    <div className="bg-zinc-900 border border-zinc-800 rounded-lg p-5">
-      <span className="text-zinc-500 text-xs font-medium uppercase tracking-wide">
-        {label}
-      </span>
-      <div className="text-white text-2xl font-bold mt-2">{value}</div>
-      {trend && (
-        <div
-          className={`flex items-center gap-1 mt-2 text-xs ${trendUp ? "text-green-400" : "text-red-400"}`}
-        >
-          {trendUp ? <TrendUpIcon /> : <TrendDownIcon />}
-          {trend}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// Simple Area Chart component
-function AreaChart({ data }: { data: { date: string; value: number }[] }) {
-  const maxValue = Math.max(...data.map((d) => d.value), 1);
-  const minValue = Math.min(...data.map((d) => d.value), 0);
-  const range = maxValue - minValue || 1;
-
-  // Generate SVG path
-  const width = 100;
-  const height = 100;
-  const points = data.map((d, i) => {
-    const x = (i / (data.length - 1)) * width;
-    const y = height - ((d.value - minValue) / range) * height;
-    return `${x},${y}`;
-  });
-
-  // Create smooth curve path
-  const pathD = `M ${points.join(" L ")}`;
-  const areaD = `${pathD} L ${width},${height} L 0,${height} Z`;
-
-  return (
-    <div className="relative h-48 w-full">
-      <svg
-        viewBox={`0 0 ${width} ${height}`}
-        className="w-full h-full"
-        preserveAspectRatio="none"
-      >
-        {/* Grid lines */}
-        {[0, 25, 50, 75, 100].map((pct) => (
-          <line
-            key={pct}
-            x1="0"
-            y1={`${pct}%`}
-            x2="100%"
-            y2={`${pct}%`}
-            stroke="rgba(255,255,255,0.05)"
-            strokeWidth="0.5"
-          />
-        ))}
-        {/* Area fill */}
-        <path d={areaD} fill="url(#blueGradient)" opacity="0.3" />
-        {/* Line */}
-        <path d={pathD} fill="none" stroke="#3B82F6" strokeWidth="2" />
-        {/* Gradient definition */}
-        <defs>
-          <linearGradient id="blueGradient" x1="0%" y1="0%" x2="0%" y2="100%">
-            <stop offset="0%" stopColor="#3B82F6" stopOpacity="0.5" />
-            <stop offset="100%" stopColor="#3B82F6" stopOpacity="0" />
-          </linearGradient>
-        </defs>
-      </svg>
-      {/* Y-axis labels */}
-      <div className="absolute left-0 top-0 bottom-0 flex flex-col justify-between text-xs text-zinc-600 -ml-2">
-        <span>{maxValue.toLocaleString()}</span>
-        <span>{Math.round((maxValue + minValue) / 2).toLocaleString()}</span>
-        <span>{minValue.toLocaleString()}</span>
-      </div>
-      {/* X-axis labels */}
-      <div className="absolute bottom-0 left-0 right-0 flex justify-between text-xs text-zinc-600 mt-2">
-        {data
-          .filter((_, i) => i % Math.ceil(data.length / 7) === 0)
-          .map((d) => (
-            <span key={d.date}>{formatShortDate(d.date)}</span>
-          ))}
-      </div>
-    </div>
-  );
-}
-
-function PieChart({
-  data,
-  total,
-}: {
-  data: { value: number; color: string }[];
-  total: number;
-}) {
-  const size = 160;
-  const strokeWidth = 18;
-  const radius = (size - strokeWidth) / 2;
-  const circumference = 2 * Math.PI * radius;
-  let offset = 0;
-
-  return (
-    <div className="relative h-40 w-40">
-      <svg viewBox={`0 0 ${size} ${size}`} className="h-full w-full">
-        <g transform={`rotate(-90 ${size / 2} ${size / 2})`}>
-          <circle
-            cx={size / 2}
-            cy={size / 2}
-            r={radius}
-            fill="transparent"
-            stroke="rgba(255,255,255,0.08)"
-            strokeWidth={strokeWidth}
-          />
-          {data.map((slice, index) => {
-            const dash = (slice.value / total) * circumference;
-            const dashArray = `${dash} ${circumference - dash}`;
-            const dashOffset = -offset;
-            offset += dash;
-            return (
-              <circle
-                key={`${slice.color}-${index}`}
-                cx={size / 2}
-                cy={size / 2}
-                r={radius}
-                fill="transparent"
-                stroke={slice.color}
-                strokeWidth={strokeWidth}
-                strokeDasharray={dashArray}
-                strokeDashoffset={dashOffset}
-              />
-            );
-          })}
-        </g>
-        <text
-          x="50%"
-          y="48%"
-          textAnchor="middle"
-          className="fill-white text-sm font-semibold"
-        >
-          {formatCost(total)}
-        </text>
-        <text
-          x="50%"
-          y="60%"
-          textAnchor="middle"
-          className="fill-zinc-500 text-[10px]"
-        >
-          Total
-        </text>
-      </svg>
-    </div>
-  );
-}
 
 export function UsageDashboard() {
   const [activeTab, setActiveTab] = useState<TabType>("overview");
@@ -386,6 +30,12 @@ export function UsageDashboard() {
   const [modelSearch, setModelSearch] = useState("");
   const [conversationSearch, setConversationSearch] = useState("");
   const [conversationStatus, setConversationStatus] = useState("running");
+
+  const [conversationSortBy, setConversationSortBy] = useState("updated_at");
+  const [conversationSortOrder, setConversationSortOrder] = useState("desc");
+  const [conversationSandboxStatus, setConversationSandboxStatus] =
+    useState("");
+
   const [conversationPage, setConversationPage] = useState(1);
   const [conversationPerPage, setConversationPerPage] = useState(20);
 
@@ -403,7 +53,10 @@ export function UsageDashboard() {
       page: conversationPage,
       perPage: conversationPerPage,
       search: conversationSearch,
+      sortBy: conversationSortBy,
+      sortOrder: conversationSortOrder,
       executionStatus: conversationStatus,
+      sandboxStatus: conversationSandboxStatus,
       timeWindow: conversationTimeWindow,
     });
 
@@ -533,13 +186,19 @@ export function UsageDashboard() {
     return organizationService.exportConversationsUrl({
       orgId: organizationId,
       search: conversationSearch || undefined,
+      sortBy: conversationSortBy,
+      sortOrder: conversationSortOrder,
       executionStatus: conversationStatus || undefined,
+      sandboxStatus: conversationSandboxStatus || undefined,
       timeWindow: conversationTimeWindow || undefined,
     });
   }, [
     organizationId,
     conversationSearch,
+    conversationSortBy,
+    conversationSortOrder,
     conversationStatus,
+    conversationSandboxStatus,
     conversationTimeWindow,
   ]);
 
@@ -607,568 +266,79 @@ export function UsageDashboard() {
       <div className="p-8">
         {/* Overview Tab */}
         {activeTab === "overview" && (
-          <div className="space-y-6">
-            {/* KPI Row */}
-            <div className="grid grid-cols-4 gap-4">
-              <KPICard
-                label="Total Conversations"
-                value={totalConversations.toLocaleString()}
-              />
-              <KPICard
-                label="Active Conversations"
-                value={activeConversations.toLocaleString()}
-              />
-              <KPICard
-                label="Avg Cost / Conversation"
-                value={`$${avgCostPerConversation.toFixed(2)}`}
-              />
-              <KPICard
-                label={`Total Spend (${timeWindowLabel})`}
-                value={totalSpend}
-              />
-            </div>
-
-            {/* Chart Row */}
-            <div className="grid grid-cols-1 gap-6 lg:grid-cols-3">
-              <div className="bg-zinc-900 border border-zinc-800 rounded-lg p-6 lg:col-span-2">
-                <div className="flex items-start justify-between mb-6">
-                  <div>
-                    <h2 className="text-lg font-medium text-white">
-                      Conversations per day
-                    </h2>
-                    <p className="text-sm text-zinc-500">
-                      {timeWindowLabel} · all users
-                    </p>
-                  </div>
-                  <button
-                    type="button"
-                    className="flex items-center gap-2 px-3 py-1.5 text-sm text-zinc-400 border border-zinc-700 rounded-lg hover:text-white hover:border-zinc-600 transition-colors"
-                  >
-                    <ExportIcon />
-                    Export CSV
-                  </button>
-                </div>
-                {chartData.length > 0 ? (
-                  <AreaChart data={chartData} />
-                ) : (
-                  <div className="py-10 text-center text-sm text-zinc-500">
-                    No usage data available yet.
-                  </div>
-                )}
-              </div>
-              <div className="bg-zinc-900 border border-zinc-800 rounded-lg p-6">
-                <div className="flex items-start justify-between mb-6">
-                  <div>
-                    <h2 className="text-lg font-medium text-white">
-                      Spend by agent
-                    </h2>
-                    <p className="text-sm text-zinc-500">
-                      {timeWindowLabel} · total spend
-                    </p>
-                  </div>
-                </div>
-                {agentSpendTotal > 0 ? (
-                  <div className="flex flex-col items-center gap-6 lg:flex-row">
-                    <PieChart
-                      data={agentSpendRows.map((row) => ({
-                        value: row.total_cost,
-                        color: row.color,
-                      }))}
-                      total={agentSpendTotal}
-                    />
-                    <div className="w-full space-y-3">
-                      {agentSpendRows.map((row) => (
-                        <div
-                          key={row.agent_name}
-                          className="flex items-center justify-between text-sm"
-                        >
-                          <div className="flex items-center gap-2">
-                            <span
-                              className="h-2 w-2 rounded-full"
-                              style={{ backgroundColor: row.color }}
-                            />
-                            <span className="text-zinc-300">
-                              {row.agent_name}
-                            </span>
-                          </div>
-                          <span className="text-zinc-400">
-                            {formatCost(row.total_cost)} ·
-                            {` ${row.percent.toFixed(1)}%`}
-                          </span>
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                ) : (
-                  <div className="py-10 text-center text-sm text-zinc-500">
-                    No agent spend data available yet.
-                  </div>
-                )}
-              </div>
-            </div>
-          </div>
+          <OverviewTab
+            totalConversations={totalConversations}
+            activeConversations={activeConversations}
+            avgCostPerConversation={avgCostPerConversation}
+            totalSpend={totalSpend}
+            timeWindowLabel={timeWindowLabel}
+            chartData={chartData}
+            agentSpendRows={agentSpendRows}
+            agentSpendTotal={agentSpendTotal}
+          />
         )}
 
         {/* Conversations Tab */}
         {activeTab === "conversations" && (
-          <div className="space-y-4">
-            {/* Filter Bar */}
-            <div className="flex items-center gap-4">
-              <div className="relative flex-1 max-w-md">
-                <span className="absolute left-3 top-1/2 -translate-y-1/2 text-zinc-500">
-                  <SearchIcon />
-                </span>
-                <input
-                  type="text"
-                  placeholder="Search by title or user..."
-                  value={conversationSearch}
-                  onChange={(e) => {
-                    setConversationSearch(e.target.value);
-                    setConversationPage(1);
-                  }}
-                  className="w-full pl-10 pr-4 py-2 bg-zinc-900 border border-zinc-800 rounded-lg text-white placeholder-zinc-500 focus:outline-none focus:border-zinc-700"
-                />
-              </div>
-              <select
-                value={conversationStatus}
-                onChange={(e) => {
-                  setConversationStatus(e.target.value);
-                  setConversationPage(1);
-                }}
-                className="px-3 py-2 bg-zinc-900 border border-zinc-800 rounded-lg text-sm text-zinc-400 focus:outline-none focus:border-zinc-700"
-              >
-                <option value="">All statuses</option>
-                <option value="running">Running</option>
-                <option value="idle">Idle</option>
-                <option value="paused">Paused</option>
-                <option value="finished">Finished</option>
-                <option value="error">Error</option>
-                <option value="stuck">Stuck</option>
-              </select>
-              <a
-                href={exportUrl}
-                className="flex items-center gap-2 px-3 py-2 text-sm text-zinc-400 border border-zinc-700 rounded-lg hover:text-white hover:border-zinc-600 transition-colors"
-              >
-                <ExportIcon />
-                Export CSV
-              </a>
-            </div>
-
-            {/* Conversations Table */}
-            <div className="bg-zinc-900 border border-zinc-800 rounded-lg overflow-hidden">
-              <table className="w-full">
-                <thead>
-                  <tr className="border-b border-zinc-800">
-                    <th className="px-4 py-3 text-left text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      User
-                    </th>
-                    <th className="px-4 py-3 text-right text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Tokens
-                    </th>
-                    <th className="px-4 py-3 text-right text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Spend
-                    </th>
-                    <th className="px-4 py-3 text-left text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Duration
-                    </th>
-                    <th className="px-4 py-3 text-left text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Started
-                    </th>
-                    <th className="px-4 py-3 text-left text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Last update
-                    </th>
-                    <th className="px-4 py-3 text-left text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Associated PR
-                    </th>
-                    <th className="px-4 py-3 text-left text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Merged?
-                    </th>
-                    <th className="px-4 py-3 text-left text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Agent
-                    </th>
-                    <th className="px-4 py-3 text-left text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Type
-                    </th>
-                    <th className="px-4 py-3 text-right text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Stop
-                    </th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {conversationsLoading && (
-                    <tr>
-                      <td
-                        colSpan={11}
-                        className="px-4 py-8 text-center text-zinc-500"
-                      >
-                        Loading conversations...
-                      </td>
-                    </tr>
-                  )}
-                  {!conversationsLoading &&
-                    (conversationsData?.items.length ?? 0) === 0 && (
-                      <tr>
-                        <td
-                          colSpan={11}
-                          className="px-4 py-8 text-center text-zinc-500"
-                        >
-                          No conversations found for this time window.
-                        </td>
-                      </tr>
-                    )}
-                  {conversationsData?.items.map((conversation) => {
-                    const isRunning =
-                      conversation.execution_status?.toLowerCase() ===
-                      "running";
-                    return (
-                      <tr
-                        key={conversation.id}
-                        className="border-b border-zinc-800/50 hover:bg-zinc-800/50 transition-colors"
-                      >
-                        <td className="px-4 py-4">
-                          <div className="text-white text-sm font-medium">
-                            {conversation.user_email?.split("@")[0] ||
-                              "Unknown"}
-                          </div>
-                          <div className="text-xs text-zinc-500">
-                            {conversation.user_email || "-"}
-                          </div>
-                        </td>
-                        <td className="px-4 py-4 text-right text-sm font-mono text-white">
-                          {formatTokens(conversation.total_tokens)}
-                        </td>
-                        <td className="px-4 py-4 text-right text-sm text-white">
-                          {formatCost(conversation.accumulated_cost)}
-                        </td>
-                        <td className="px-4 py-4 text-sm text-zinc-400">
-                          {formatDuration(
-                            conversation.created_at,
-                            conversation.updated_at,
-                          )}
-                        </td>
-                        <td className="px-4 py-4 text-sm text-zinc-400">
-                          {formatDateTimeOrDash(conversation.created_at)}
-                        </td>
-                        <td className="px-4 py-4 text-sm text-zinc-400">
-                          {formatDateTimeOrDash(conversation.updated_at)}
-                        </td>
-                        <td className="px-4 py-4 text-sm text-zinc-400">
-                          {formatAssociatedPr(conversation)}
-                        </td>
-                        <td className="px-4 py-4 text-sm text-zinc-400">
-                          {formatMergedStatus(conversation.pr_merged)}
-                        </td>
-                        <td className="px-4 py-4 text-sm text-zinc-400">
-                          {formatAgentLabel(conversation)}
-                        </td>
-                        <td className="px-4 py-4 text-sm text-zinc-400 capitalize">
-                          {conversation.trigger || "-"}
-                        </td>
-                        <td className="px-4 py-4 text-right text-sm">
-                          {isRunning && (
-                            <button
-                              type="button"
-                              onClick={() =>
-                                handleStop({
-                                  id: conversation.id,
-                                  title: conversation.title ?? null,
-                                })
-                              }
-                              disabled={stoppingIds.has(conversation.id)}
-                              className="inline-flex items-center gap-1.5 px-2 py-1 text-xs text-zinc-400 hover:text-white hover:bg-zinc-800 rounded transition-colors disabled:opacity-40 disabled:hover:bg-transparent disabled:hover:text-zinc-400 disabled:cursor-not-allowed"
-                              title="Stop conversation"
-                              aria-label="Stop conversation"
-                            >
-                              <StopIcon />
-                              {stoppingIds.has(conversation.id)
-                                ? "Stopping…"
-                                : "Stop"}
-                            </button>
-                          )}
-                        </td>
-                      </tr>
-                    );
-                  })}
-                </tbody>
-              </table>
-              <div className="flex items-center justify-between px-4 py-3 border-t border-zinc-800">
-                <div className="flex items-center gap-2">
-                  <button
-                    type="button"
-                    onClick={() =>
-                      setConversationPage((prev) => Math.max(1, prev - 1))
-                    }
-                    disabled={conversationPage <= 1}
-                    className={`flex items-center gap-1 px-2 py-1 text-sm rounded transition-colors ${
-                      conversationPage <= 1
-                        ? "text-zinc-600 cursor-not-allowed"
-                        : "text-zinc-400 hover:text-white hover:bg-zinc-800"
-                    }`}
-                  >
-                    Previous
-                  </button>
-                  <button
-                    type="button"
-                    onClick={() =>
-                      setConversationPage((prev) =>
-                        Math.min(conversationTotalPages, prev + 1),
-                      )
-                    }
-                    disabled={conversationPage >= conversationTotalPages}
-                    className={`flex items-center gap-1 px-2 py-1 text-sm rounded transition-colors ${
-                      conversationPage >= conversationTotalPages
-                        ? "text-zinc-600 cursor-not-allowed"
-                        : "text-zinc-400 hover:text-white hover:bg-zinc-800"
-                    }`}
-                  >
-                    Next
-                  </button>
-                </div>
-                <div className="flex items-center gap-4">
-                  <div className="flex items-center gap-2">
-                    <span className="text-zinc-500 text-sm">Per page</span>
-                    <select
-                      value={conversationPerPage}
-                      onChange={(e) => {
-                        setConversationPerPage(Number(e.target.value));
-                        setConversationPage(1);
-                      }}
-                      className="px-2 py-1 bg-zinc-900 border border-zinc-800 rounded text-sm text-white focus:outline-none"
-                    >
-                      <option value="10">10</option>
-                      <option value="20">20</option>
-                      <option value="50">50</option>
-                    </select>
-                  </div>
-                  <span className="text-zinc-500 text-sm">
-                    Page {conversationPage} of {conversationTotalPages} ·{" "}
-                    {conversationTotalItems} conversations
-                  </span>
-                </div>
-              </div>
-            </div>
-
-            {pendingStop && (
-              <ConfirmationModal
-                text={stopConfirmationText}
-                onConfirm={confirmStop}
-                onCancel={cancelStop}
-              />
-            )}
-          </div>
+          <ConversationsTab
+            conversationSearch={conversationSearch}
+            conversationStatus={conversationStatus}
+            conversationSortBy={conversationSortBy}
+            conversationSortOrder={conversationSortOrder}
+            conversationSandboxStatus={conversationSandboxStatus}
+            exportUrl={exportUrl}
+            conversationPage={conversationPage}
+            conversationPerPage={conversationPerPage}
+            conversationTotalPages={conversationTotalPages}
+            conversationTotalItems={conversationTotalItems}
+            conversationsLoading={conversationsLoading}
+            conversationsData={conversationsData}
+            stoppingIds={stoppingIds}
+            onSearchChange={(value) => {
+              setConversationSearch(value);
+              setConversationPage(1);
+            }}
+            onStatusChange={(value) => {
+              setConversationStatus(value);
+              setConversationPage(1);
+            }}
+            onSortByChange={(value) => {
+              setConversationSortBy(value);
+              setConversationPage(1);
+            }}
+            onSortOrderChange={(value) => {
+              setConversationSortOrder(value);
+              setConversationPage(1);
+            }}
+            onSandboxStatusChange={(value) => {
+              setConversationSandboxStatus(value);
+              setConversationPage(1);
+            }}
+            onPageChange={setConversationPage}
+            onPerPageChange={(value) => {
+              setConversationPerPage(value);
+              setConversationPage(1);
+            }}
+            onStopConversation={handleStop}
+            pendingStop={pendingStop}
+            stopConfirmationText={stopConfirmationText}
+            onConfirmStop={confirmStop}
+            onCancelStop={cancelStop}
+          />
         )}
 
         {/* Users Tab */}
         {activeTab === "users" && (
-          <div className="space-y-4">
-            <div className="bg-zinc-900 border border-zinc-800 rounded-lg overflow-hidden">
-              <table className="w-full">
-                <thead>
-                  <tr className="border-b border-zinc-800">
-                    <th className="px-4 py-3 text-left text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      User
-                    </th>
-                    <th className="px-4 py-3 text-right text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Convos
-                    </th>
-                    <th className="px-4 py-3 text-left text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      First convo
-                    </th>
-                    <th className="px-4 py-3 text-left text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Last convo
-                    </th>
-                    <th className="px-4 py-3 text-left text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      First login
-                    </th>
-                    <th className="px-4 py-3 text-left text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Last login
-                    </th>
-                    <th className="px-4 py-3 text-right text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Spend MTD
-                    </th>
-                    <th className="px-4 py-3 text-right text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Spend YTD
-                    </th>
-                    <th className="px-4 py-3 text-right text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Lifetime
-                    </th>
-                    <th className="px-4 py-3 text-left text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Budget
-                    </th>
-                    <th className="px-4 py-3 text-right text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      PRs merged
-                    </th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {userUsageLoading && (
-                    <tr>
-                      <td
-                        colSpan={11}
-                        className="px-4 py-8 text-center text-zinc-500"
-                      >
-                        Loading user usage...
-                      </td>
-                    </tr>
-                  )}
-                  {!userUsageLoading &&
-                    (userUsage?.items.length ?? 0) === 0 && (
-                      <tr>
-                        <td
-                          colSpan={11}
-                          className="px-4 py-8 text-center text-zinc-500"
-                        >
-                          No user usage data available yet.
-                        </td>
-                      </tr>
-                    )}
-                  {userUsage?.items.map((user) => (
-                    <tr
-                      key={user.user_id}
-                      className="border-b border-zinc-800/50 hover:bg-zinc-800/50 transition-colors"
-                    >
-                      <td className="px-4 py-4">
-                        <div className="text-white text-sm font-medium">
-                          {user.user_name ??
-                            user.user_email?.split("@")[0] ??
-                            "Unknown"}
-                        </div>
-                        <div className="text-xs text-zinc-500">
-                          {user.user_email || "-"}
-                        </div>
-                      </td>
-                      <td className="px-4 py-4 text-right text-sm text-white">
-                        {user.conversation_count.toLocaleString()}
-                      </td>
-                      <td className="px-4 py-4 text-sm text-zinc-400">
-                        {formatDateTimeOrDash(user.first_conversation_at)}
-                      </td>
-                      <td className="px-4 py-4 text-sm text-zinc-400">
-                        {formatDateTimeOrDash(user.last_conversation_at)}
-                      </td>
-                      <td className="px-4 py-4 text-sm text-zinc-400">
-                        {formatDateTimeOrDash(user.first_login_at)}
-                      </td>
-                      <td className="px-4 py-4 text-sm text-zinc-400">
-                        {formatDateTimeOrDash(user.last_login_at)}
-                      </td>
-                      <td className="px-4 py-4 text-right text-sm text-white">
-                        {formatCost(user.spend_mtd)}
-                      </td>
-                      <td className="px-4 py-4 text-right text-sm text-white">
-                        {formatCost(user.spend_ytd)}
-                      </td>
-                      <td className="px-4 py-4 text-right text-sm text-white">
-                        {formatCost(user.spend_lifetime)}
-                      </td>
-                      <td className="px-4 py-4 text-sm text-zinc-400">
-                        {formatBudget(user)}
-                      </td>
-                      <td className="px-4 py-4 text-right text-sm text-zinc-400">
-                        {user.prs_merged ?? "-"}
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </div>
-          </div>
+          <UsersTab userUsage={userUsage} userUsageLoading={userUsageLoading} />
         )}
 
         {/* Models Tab */}
         {activeTab === "models" && (
-          <div className="space-y-4">
-            {/* Filter Bar */}
-            <div className="flex items-center justify-between">
-              <div className="relative w-64">
-                <span className="absolute left-3 top-1/2 -translate-y-1/2 text-zinc-500">
-                  <SearchIcon />
-                </span>
-                <input
-                  type="text"
-                  placeholder="Search models..."
-                  value={modelSearch}
-                  onChange={(e) => setModelSearch(e.target.value)}
-                  className="w-full pl-10 pr-4 py-2 bg-zinc-900 border border-zinc-800 rounded-lg text-white placeholder-zinc-500 focus:outline-none focus:border-zinc-700"
-                />
-              </div>
-              <button
-                type="button"
-                className="flex items-center gap-2 px-3 py-2 text-sm text-zinc-400 border border-zinc-700 rounded-lg hover:text-white hover:border-zinc-600 transition-colors"
-              >
-                <ExportIcon />
-                Export CSV
-              </button>
-            </div>
-
-            {/* Models Table */}
-            <div className="bg-zinc-900 border border-zinc-800 rounded-lg overflow-hidden">
-              <table className="w-full">
-                <thead>
-                  <tr className="border-b border-zinc-800">
-                    <th className="px-4 py-3 text-left text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Model
-                    </th>
-                    <th className="px-4 py-3 text-right text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Conversations
-                    </th>
-                    <th className="px-4 py-3 text-right text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Tokens Used
-                    </th>
-                    <th className="px-4 py-3 text-right text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Avg Tokens / Convo
-                    </th>
-                    <th className="px-4 py-3 text-right text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Avg Cost / Convo
-                    </th>
-                    <th className="px-4 py-3 text-right text-xs font-medium text-zinc-500 uppercase tracking-wider">
-                      Total Cost
-                    </th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {filteredModels.map((model) => (
-                    <tr
-                      key={model.model_name}
-                      className="border-b border-zinc-800/50 hover:bg-zinc-800/50 transition-colors"
-                    >
-                      <td className="px-4 py-5">
-                        <div className="text-white font-medium">
-                          {model.model_name}
-                        </div>
-                      </td>
-                      <td className="px-4 py-5 text-white text-sm font-mono text-right">
-                        {model.conversation_count.toLocaleString()}
-                      </td>
-                      <td className="px-4 py-5 text-white text-sm font-mono text-right">
-                        {formatTokens(model.total_tokens)}
-                      </td>
-                      <td className="px-4 py-5 text-white text-sm font-mono text-right">
-                        {formatTokens(model.avgTokens)}
-                      </td>
-                      <td className="px-4 py-5 text-white text-sm font-mono text-right">
-                        ${model.avgCost.toFixed(2)}
-                      </td>
-                      <td className="px-4 py-5 text-white text-sm font-mono text-right font-medium">
-                        ${model.total_cost.toFixed(2)}
-                      </td>
-                    </tr>
-                  ))}
-
-                  {filteredModels.length === 0 && (
-                    <tr>
-                      <td
-                        colSpan={6}
-                        className="px-4 py-8 text-center text-zinc-500"
-                      >
-                        No model usage data available for this time window.
-                      </td>
-                    </tr>
-                  )}
-                </tbody>
-              </table>
-            </div>
-          </div>
+          <ModelsTab
+            modelSearch={modelSearch}
+            onModelSearchChange={setModelSearch}
+            filteredModels={filteredModels}
+          />
         )}
       </div>
     </div>
